backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
redis==5.0.1
pillow==10.1.0
opencv-python==4.8.1.78
httpx==0.25.2
pytest==7.4.3
pinecone-client==2.2.4
numpy==1.24.3
